[ORBI-14] fixed long conversion

+++ Orbi-Server/src/main/java/foundation/softwaredesign/orbi/service/game/character/CharacterService.java
@@ -113,8 +113,8 @@ public class CharacterService {
         if (isNull(currentCharacter.getXp()) || (nonNull(currentCharacter.getXp()) && (currentCharacter.getXp().longValue() == 0))) {
             currentCharacter.setLevel(new Long(0));
         }
-        int level = (int)((int)Math.log(currentCharacter.getXp()) / Math.log(16));
-        currentCharacter.setLevel(new Long(level + 1));
+        long level = (long)((long)Math.log(currentCharacter.getXp()) / Math.log(16));
+        currentCharacter.setLevel(new Long(level + 1l));
     }
 
     public void gifted() {
